api/markopolo/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httpie==2.4.0
idna==2.10
Pygments==2.8.1
PyJWT==2.0.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.3
